requirements.txt
praw
spotipy
pandas
python-dotenv